+++ metadata.rb
@@ -1,9 +1,9 @@
 # Encoding: utf-8
-name 'ICanHazDevops'
+name 'icanhazdevops'
 maintainer 'Helmut Knoepfle'
 maintainer_email 'none'
 license 'MIT'
-description 'Installs/Configures ICanHazDevops'
+description 'Installs/Configures icanhazdevops'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version '0.2.0'
+version '0.2.1'
 depends 'apache2'