+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-ICanHazDevops CHANGELOG
+icanhazdevops CHANGELOG
 =======================
 
-This file is used to list changes made in each version of the ICanHazDevops cookbook.
+This file is used to list changes made in each version of the icanhazdevops cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of ICanHazDevops
+- [your_name] - Initial release of icanhazdevops
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.